Automatically Committed for MakeGamesWithUs

+++ GameOfLife.spritebuilder/Grid.m
@@ -103,7 +103,7 @@ static const int GRID_COLUMNS = 10;
     for (int i = 0; i < [_gridArray count]; i++) {
         for (int j = 0; j < [_gridArray[i] count]; j++) {
             Creature *currentCreature = _gridArray[i][j];
-            int z = currentCreature.livingNeighbors;
+            long z = currentCreature.livingNeighbors;
             if (z == 3) {
                 currentCreature.isAlive = true;
                 numAlive++;
